Rename project from kc2spicedb to k2spicedb and update related files; add CLI tool for Keycloak to SpiceDB migration
