+++ k2spicedb/schema_generator.py
