+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "kc2spicedb"
+name = "k2spicedb"
 version = "0.1.0"
 description = "A tool to migrate Keycloak realms to SpiceDB schemas"
 authors = ["James Staud <james@example.com>"]
